Migrate bonsai-sdk to maybe_async (#1983)
This PR attempts to replace the old async module of bonsai-sdk with the
maybe-async (https://crates.io/crates/maybe_async) method of proc-macro
rewriting.

---------
