+++ Cargo.toml
@@ -34,7 +34,7 @@ homepage = "https://risczero.com/"
 repository = "https://github.com/risc0/risc0/"
 
 [workspace.dependencies]
-bonsai-sdk = { version = "0.8.0", default-features = false, path = "bonsai/sdk" }
+bonsai-sdk = { version = "0.9.0-alpha.1", default-features = false, path = "bonsai/sdk" }
 hotbench = { path = "tools/hotbench" }
 risc0-binfmt = { version = "1.0.1", default-features = false, path = "risc0/binfmt" }
 risc0-build = { version = "1.0.1", default-features = false, path = "risc0/build" }
